+++ authress_sdk/models/access_record_group.py
@@ -1,0 +1,100 @@
+# coding: utf-8
+
+
+import pprint
+import re
+
+import six
+
+
+class AccessRecordGroup(object):
+
+    """
+    Attributes:
+      swagger_types (dict): The key is attribute name
+                            and the value is attribute type.
+      attribute_map (dict): The key is attribute name
+                            and the value is json key in definition.
+    """
+    swagger_types = {
+        'group_id': 'str'
+    }
+
+    attribute_map = {
+        'group_id': 'groupId'
+    }
+
+    def __init__(self, group_id=None):
+        """AccessRecordGroup"""
+        self._group_id = None
+        self.discriminator = None
+        self.group_id = group_id
+
+    @property
+    def group_id(self):
+        """Gets the group_id of this AccessRecordGroup.
+
+
+        :return: The group_id of this AccessRecordGroup.
+        :rtype: str
+        """
+        return self._group_id
+
+    @group_id.setter
+    def group_id(self, group_id):
+        """Sets the group_id of this AccessRecordGroup.
+
+
+        :param group_id: The group_id of this AccessRecordGroup.
+        :type: str
+        """
+        if group_id is None:
+            raise ValueError("Invalid value for `group_id`, must not be `None`")
+
+        self._group_id = group_id
+
+    def to_dict(self):
+        """Returns the model properties as a dict"""
+        result = {}
+
+        for attr, _ in six.iteritems(self.swagger_types):
+            value = getattr(self, attr)
+            if isinstance(value, list):
+                result[attr] = list(map(
+                    lambda x: x.to_dict() if hasattr(x, "to_dict") else x,
+                    value
+                ))
+            elif hasattr(value, "to_dict"):
+                result[attr] = value.to_dict()
+            elif isinstance(value, dict):
+                result[attr] = dict(map(
+                    lambda item: (item[0], item[1].to_dict())
+                    if hasattr(item[1], "to_dict") else item,
+                    value.items()
+                ))
+            else:
+                result[attr] = value
+        if issubclass(AccessRecordGroup, dict):
+            for key, value in self.items():
+                result[key] = value
+
+        return result
+
+    def to_str(self):
+        """Returns the string representation of the model"""
+        return pprint.pformat(self.to_dict())
+
+    def __repr__(self):
+        """For `print` and `pprint`"""
+        return self.to_str()
+
+    def __eq__(self, other):
+        """Returns true if both objects are equal"""
+        if not isinstance(other, AccessRecordGroup):
+            return False
+
+        return self.__dict__ == other.__dict__
+
+    def __ne__(self, other):
+        """Returns true if both objects are not equal"""
+        return not self == other