+++ CHANGELOG.md
@@ -3,3 +3,4 @@ This is the changelog for [Authress SDK](readme.md).
 
 ## 1.0 ##
 * Add EdDSA support
+* Add `Groups` to `AccessRecords`